+++ R/newS.R
@@ -1,6 +1,6 @@
 newS <- function(y) {
   ind <- order(y[,1])
-  list(time = as.numeric(y[ind,1]),
-       fail = as.numeric(y[ind,2]),
+  list(time = as.double(y[ind,1]),
+       fail = as.double(y[ind,2]),
        ind  = ind)
 }
